cssazurelab/requirements.txt
azure.cli.core
secrets